refactor(fsd): isolate view-specific organisms into features
