+++ eslint.config.mjs
@@ -193,15 +193,15 @@ export default ts.config(
       "src/components/molecules/DeleteConfirmModal.tsx",
       "src/components/organisms/HistoryTab.tsx",
       "src/components/organisms/SettingsTab.tsx",
-      "src/components/organisms/MintingView.tsx",
-      "src/components/organisms/CardDetailView.tsx",
+      "src/features/minting/components/MintingView.tsx",
+      "src/features/card-detail/components/CardDetailView.tsx",
       "src/components/organisms/Workbench.tsx",
       "src/components/organisms/InteractiveTutorial.tsx",
       "src/components/organisms/WebLlmSettingsSection.tsx",
       "src/components/organisms/SimpleWorkbenchModal.tsx",
       "src/components/organisms/SlotVariablesSection.tsx",
       "src/components/organisms/StorageManagerSection.tsx",
-      "src/components/organisms/CategoryManagerModal.tsx",
+      "src/features/category-manager/components/CategoryManagerModal.tsx",
       "src/components/organisms/CloudSyncSection.tsx",
       "src/components/organisms/LocalBackupSection.tsx",
       "src/components/organisms/Cauldron.tsx",
@@ -308,14 +308,14 @@ export default ts.config(
       "src/components/organisms/AiRecipeAdviceSection.tsx",
       "src/components/organisms/AiStyleAnalysisSection.tsx",
       "src/components/organisms/BrandLogoSettingsSection.tsx",
-      "src/components/organisms/CardDetailView/SubSections.tsx",
+      "src/features/card-detail/components/CardDetailView/SubSections.tsx",
       "src/components/organisms/CardsGrid.tsx",
       "src/components/organisms/CategoryFilters.tsx",
-      "src/components/organisms/CategoryManagerModal/CardSelectionView.tsx",
-      "src/components/organisms/CategoryManagerModal/CategoryForm.tsx",
-      "src/components/organisms/CategoryManagerModal/CategoryList.tsx",
-      "src/components/organisms/CategoryManagerModal/CategoryModalHeader.tsx",
-      "src/components/organisms/CategoryManagerModal/CoverSettingsFields.tsx",
+      "src/features/category-manager/components/CategoryManagerModal/CardSelectionView.tsx",
+      "src/features/category-manager/components/CategoryManagerModal/CategoryForm.tsx",
+      "src/features/category-manager/components/CategoryManagerModal/CategoryList.tsx",
+      "src/features/category-manager/components/CategoryManagerModal/CategoryModalHeader.tsx",
+      "src/features/category-manager/components/CategoryManagerModal/CoverSettingsFields.tsx",
       "src/components/organisms/CloudSyncSection.tsx",
       "src/components/organisms/DangerZoneSection.tsx",
       "src/components/organisms/DbErrorOverlay.tsx",
@@ -329,9 +329,9 @@ export default ts.config(
       "src/components/organisms/LibrarySearchBar.tsx",
       "src/components/organisms/LocalBackupSection.tsx",
       "src/components/organisms/MergeStackMaterialList.tsx",
-      "src/components/organisms/MintingView.tsx",
-      "src/components/organisms/MintingView/CardIdentitySubSections.tsx",
-      "src/components/organisms/MintingView/MintingViewContent.tsx",
+      "src/features/minting/components/MintingView.tsx",
+      "src/features/minting/components/MintingView/CardIdentitySubSections.tsx",
+      "src/features/minting/components/MintingView/MintingViewContent.tsx",
       "src/components/organisms/NonTargetSiteView.tsx",
       "src/components/organisms/OnboardingGuide.tsx",
       "src/components/organisms/OnboardingGuideComponents.tsx",
