+++ memory-bank/userJourneys.json
@@ -7,7 +7,7 @@
       "steps": ["画像上のMintボタン押下", "プロンプト解析", "カード保存"],
       "nextJourneys": ["@J-ORG-EXPERT-01", "@J-WB-EXPERT-01"],
       "viewComponents": [
-        "src/components/organisms/MintingView.tsx",
+        "src/features/minting/components/MintingView.tsx",
         "src/lib/mj-parser.ts"
       ]
     },
@@ -61,7 +61,9 @@
         "カードに割り当て"
       ],
       "nextJourneys": ["@J-ORG-EXPERT-01", "@J-ORG-FOLDER-01"],
-      "viewComponents": ["src/components/organisms/CategoryManagerModal.tsx"]
+      "viewComponents": [
+        "src/features/category-manager/components/CategoryManagerModal.tsx"
+      ]
     },
     {
       "id": "@J-ORG-EASY-01",
@@ -324,7 +326,7 @@
         "カード保存"
       ],
       "nextJourneys": ["@J-ORG-EXPERT-01"],
-      "viewComponents": ["src/components/organisms/MintingView.tsx"]
+      "viewComponents": ["src/features/minting/components/MintingView.tsx"]
     },
     {
       "id": "@J-ORG-VERSION-01",
@@ -338,7 +340,9 @@
         "保存して確定"
       ],
       "nextJourneys": ["@J-ORG-EXPERT-01"],
-      "viewComponents": ["src/components/organisms/CardDetailView.tsx"]
+      "viewComponents": [
+        "src/features/card-detail/components/CardDetailView.tsx"
+      ]
     },
     {
       "id": "@J-WB-MIXING-WEIGHTS-01",
@@ -587,9 +591,9 @@
       "nextJourneys": ["@J-ORG-EXPERT-01"],
       "viewComponents": [
         "src/components/organisms/LibraryTab.tsx",
-        "src/components/organisms/CategoryManagerModal.tsx",
-        "src/components/organisms/CategoryManagerModal/CategoryForm.tsx",
-        "src/components/organisms/CategoryManagerModal/CoverSettingsFields.tsx",
+        "src/features/category-manager/components/CategoryManagerModal.tsx",
+        "src/features/category-manager/components/CategoryManagerModal/CategoryForm.tsx",
+        "src/features/category-manager/components/CategoryManagerModal/CoverSettingsFields.tsx",
         "src/components/organisms/FolderExplorer.tsx",
         "src/lib/image-utils.ts"
       ]
