+++ src/components/organisms/LibraryTab.tsx
@@ -4,6 +4,7 @@ import React, { useState } from "react"
 import { useLanguage } from "../../contexts/LanguageContext"
 import { useSettings } from "../../contexts/SettingsContext"
 import { useTutorial } from "../../contexts/TutorialContext"
+import { CategoryManagerModal } from "../../features/category-manager/components/CategoryManagerModal"
 import { useLibrary } from "../../hooks/useLibrary"
 import { useWebLlm } from "../../hooks/useWebLlm"
 import { THEME_STYLES } from "../../lib/theme-config"
@@ -12,7 +13,6 @@ import { OpfsImage } from "../atoms/OpfsImage"
 import { type AlertType } from "../molecules/ConnectionAlert"
 import { LocalAiSetupPlaceholder } from "../molecules/LocalAiSetupPlaceholder"
 import { CardsGrid } from "./CardsGrid"
-import { CategoryManagerModal } from "./CategoryManagerModal"
 import { EmptyState } from "./EmptyState"
 import { FolderExplorer } from "./FolderExplorer"
 import { SearchAndFilterSection } from "./LibraryTab/SearchAndFilterSection"
