+++ src/features/card-detail/components/CardDetailView.tsx
