+++ src/features/card-detail/components/CardDetailView/ParametersSection.tsx
