+++ src/features/category-manager/components/CategoryManagerModal/CardSelectionView.tsx
