+++ src/features/category-manager/components/CategoryManagerModal/CoverSettingsFields.tsx
