+++ src/features/minting/components/MintingView.tsx
