+++ src/features/minting/components/MintingView/MintingViewContent.tsx
